fixed large offset

+++ src/kseq.h
@@ -44,11 +44,11 @@ typedef struct {
 kstream_t *ks_init(gzFile f);
 void ks_destroy(kstream_t *ks);
 int ks_getc(kstream_t *ks);
-int ks_getuntil2(kstream_t *ks, int delimiter, kstring_t *str, int *dret, int append);
-int ks_getuntil(kstream_t *ks, int delimiter, kstring_t *str, int *dret);
+Py_ssize_t ks_getuntil2(kstream_t *ks, int delimiter, kstring_t *str, int *dret, int append);
+Py_ssize_t ks_getuntil(kstream_t *ks, int delimiter, kstring_t *str, int *dret);
 kseq_t *kseq_init(gzFile fd);
 void kseq_rewind(kseq_t *ks);
 void kseq_destroy(kseq_t *ks);
-int kseq_read(kseq_t *seq);
+Py_ssize_t kseq_read(kseq_t *seq);
 
 #endif
